solana_zk_sdk::zk_elgamal_proof_program::proof_data

Module ciphertext_ciphertext_equality

The ciphertext-ciphertext equality proof instruction.

A ciphertext-ciphertext equality proof is defined with respect to two twisted ElGamal ciphertexts. The proof certifies that the two ciphertexts encrypt the same message. To generate the proof, a prover must provide the decryption key for the first ciphertext and the randomness used to generate the second ciphertext.

Structs

CiphertextCiphertextEqualityProofContext

The context data needed to verify a ciphertext-ciphertext equality proof.

CiphertextCiphertextEqualityProofData

The instruction data that is needed for the ProofInstruction::VerifyCiphertextCiphertextEquality instruction.